% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts.R, R/hts.R
\name{get_groups}
\alias{get_groups}
\alias{get_nodes}
\title{Get nodes/groups from an hts/gts object}
\usage{
get_groups(y)

get_nodes(y)
}
\arguments{
\item{y}{An hts or gts object
series.}
}
\description{
Get nodes/groups from an hts/gts object
}
