\name{htseg1}
\Rdversion{1.1}
\alias{htseg1}
\docType{data}
\title{
Simulated data sets
}
\description{
We simulate random data following a standard normal distribution. The simulated data have 4 levels. 
}
\usage{data(htseg1)}
\format{
An object of class \code{\link[hts]{hts}}.
}
\references{
R. J. Hyndman, R. A. Ahmed, G. Athanasopoulos and H.L. Shang (2010) 
Optimal combination forecasts for hierarchical time series. 
Working paper, Monash University. 
\url{http://robjhyndman.com/papers/hierarchical}
}
\examples{
plot(htseg1)
}
\keyword{datasets}
