\name{htseg2}
\Rdversion{1.1}
\alias{htseg2}
\docType{data}
\title{
Simulated data sets
}
\description{
We simulate random data following a standard normal distribution. The simulated data have 3 levels. 
}
\usage{data(htseg2)}
\format{
An object of class \code{\link[hts]{hts}}.
}
\references{
R. J. Hyndman, R. A. Ahmed, G. Athanasopoulos and H.L. Shang (2011) 
Optimal combination forecasts for hierarchical time series. 
\emph{Computational Statistics and Data Analysis}, \bold{55}(9), 2579--2589.
\url{http://robjhyndman.com/papers/hierarchical/}
}
\examples{
plot(htseg2)
}
\keyword{datasets}
