% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts-package.R
\docType{data}
\name{infantgts}
\alias{infantgts}
\title{Regional infant mortality counts across Australia from 1933 to 2003.}
\format{
Objects of class \code{\link[hts]{gts}}.
}
\description{
These are infant mortality counts. This data set is an example of
\code{gts}, where the total infant mortality count in Australia can be first
disaggregated by sex then by state, or vice versa.
}
\examples{

plot(infantgts)

}
\references{
R. J. Hyndman, R. A. Ahmed, G. Athanasopoulos and H.L. Shang
(2011) Optimal combination forecasts for hierarchical time series.
\emph{Computational Statistics and Data Analysis}, \bold{55}(9), 2579--2589.
}
\keyword{datasets}
