\name{simulatedata4}
\Rdversion{1.1}
\alias{simulatedata4}
\docType{data}
\title{
Simulated data sets
}
\description{
We simulate random data following a standard normal distribution. The simulated data have 1 level. 
}
\usage{data(simulatedata4)}
\format{
An object of class \code{\link[hts]{hts}}.
}
\references{
G. Athanasopoulos, R. A. Ahmed and R. J. Hyndman (2009) Hierarchical forecasts for Australian domestic tourism, \emph{International Journal of Forecasting}, \bold{25}, 146-166.

R. J. Hyndman, R. A. Ahmed and G. Athanasopoulos (2007) Optimal combination forecasts for hierarchical time series. Working paper. 
Monash University Business and Economic Forecasting Unit.
\url{http://robjhyndman.com/papers/hierarchical}
}
\examples{
plot(simulatedata4)
}
\keyword{datasets}
