% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggts.R
\name{allts}
\alias{allts}
\title{Extract all time series from a gts object}
\usage{
allts(y, forecasts = TRUE)
}
\arguments{
\item{y}{An object of class \code{\link[hts]{gts}}.}

\item{forecasts}{If \code{y} contains forecasts and historical data, then
\code{forecasts} indicates whether to return the forecasts or the historical
data. Otherwise it is ignored.}
}
\description{
The time series from all levels of a hierarchical/grouped time series or a
forecasted hierarchical/grouped time series are returned as a multivariate
time series.
}
\examples{

allts(htseg1)

}
\seealso{
\code{\link[hts]{aggts}}
}
\author{
Rob J Hyndman
}
\keyword{ts}
