\name{hts-package}
\alias{hts-package}
\docType{package}
\title{
Hierarchical and grouped time series 
}
\description{
This package presents functions to create, plot and forecast hierarchical and grouped time series. In forecasting hierarchical and grouped time series, the base methods implemented include ETS, ARIMA and the naive (random walk) models. Forecasts for grouped time series are calibrated using bottom-up and optimal combination methods. Forecasts for hierarchical time series are distributed in the hierarchy using bottom-up, top-down, middle-out and optimal combination methods. Three top-down methods are available: the two Gross-Sohl methods and the forecast-proportion approach of Hyndman, Ahmed, and Athanasopoulos (2011).
}
\author{
Rob J Hyndman, Earo Wang and Alan Lee with contributions from Roman A Ahmed and Han Lin Shang to earlier versions of the package

Maintainer: Rob J Hyndman <Rob.Hyndman@monash.edu>
}
\references{
G. Athanasopoulos, R. A. Ahmed and R. J. Hyndman (2009) Hierarchical forecasts for Australian domestic tourism, \emph{International Journal of Forecasting}, \bold{25}, 146-166.

R. J. Hyndman, R. A. Ahmed, G. Athanasopoulos and H.L. Shang (2011) 
Optimal combination forecasts for hierarchical time series. 
\emph{Computational Statistics and Data Analysis}, \bold{55}(9), 2579--2589.
\url{http://robjhyndman.com/papers/hierarchical/}


Hyndman, R. J., Lee, A., & Wang, E. (2014). 
Fast computation of reconciled forecasts for hierarchical and grouped time series. 
\emph{Working paper 17/14, Department of Econometrics & Business Statistics, Monash University.}
\url{http://robjhyndman.com/working-papers/hgts/}
}
\keyword{package}


