% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts-package.R
\docType{data}
\name{htseg1}
\alias{htseg1}
\alias{htseg2}
\title{Simple examples of hierarchical time series.}
\format{
Objects of class \code{\link[hts]{hts}}.
}
\description{
These are simulated data. \code{htseg1} has three levels with a total of 8
series each of length 10. \code{htseg2} has four levels with a total of 17
series each of length 16.
}
\examples{

plot(htseg1)

}
\references{
R. J. Hyndman, R. A. Ahmed, G. Athanasopoulos and H.L. Shang
(2011) Optimal combination forecasts for hierarchical time series.
\emph{Computational Statistics and Data Analysis}, \bold{55}(9), 2579--2589.
\url{http://robjhyndman.com/papers/hierarchical/}
}
\keyword{datasets}
