% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window-gts.R
\name{window.gts}
\alias{window.gts}
\title{Time window of a gts object}
\usage{
\method{window}{gts}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link[hts]{gts}}.}

\item{...}{All other arguments are passed to \code{\link[stats]{window.ts}}.}
}
\description{
Extracts a subset of the time series from a grouped time series object.
}
\examples{

window(htseg2, start = 2000, end = 2001)

}
\author{
Rob J Hyndman
}
\keyword{ts}
